refactor compress

+++ include/common/tcol.h
@@ -86,11 +86,11 @@ bool checkColumnCompressOrSetDefault(uint8_t type, char compress[TSDB_CL_COMPRES
 bool checkColumnLevel(char level[TSDB_CL_COMPRESS_OPTION_LEN]);
 bool checkColumnLevelOrSetDefault(uint8_t type, char level[TSDB_CL_COMPRESS_OPTION_LEN]);
 
-void   setColEncode(uint32_t* compress, uint8_t encode);
-void   setColCompress(uint32_t* compress, uint16_t compressType);
-void   setColLevel(uint32_t* compress, uint8_t level);
-int8_t setColCompressByOption(uint8_t type, uint8_t encode, uint16_t compressType, uint8_t level, bool check,
-                              uint32_t* compress);
+void    setColEncode(uint32_t* compress, uint8_t encode);
+void    setColCompress(uint32_t* compress, uint16_t compressType);
+void    setColLevel(uint32_t* compress, uint8_t level);
+int32_t setColCompressByOption(uint8_t type, uint8_t encode, uint16_t compressType, uint8_t level, bool check,
+                               uint32_t* compress);
 
 int8_t validColCompressLevel(uint8_t type, uint8_t level);
 int8_t validColCompress(uint8_t type, uint8_t l2);
